@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --poll=2000",
+    "start": "ng serve --poll=2000 --port 4201",
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
@@ -12,51 +12,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.0",
-    "@angular/common": "^5.1.0",
-    "@angular/compiler": "^5.1.0",
-    "@angular/core": "^5.1.0",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "^5.1.0",
-    "@angular/platform-browser": "^5.1.0",
-    "@angular/platform-browser-dynamic": "^5.1.0",
-    "@angular/router": "^5.1.0",
+    "@angular/animations": "^7.2.0",
+    "@angular/cli": "^7.2.0",
+    "@angular/common": "^7.2.0",
+    "@angular/compiler": "^7.2.0",
+    "@angular/core": "^7.2.0",
+    "@angular/forms": "^7.2.0",
+    "@angular/http": "^7.2.0",
+    "@angular/platform-browser": "^7.2.0",
+    "@angular/platform-browser-dynamic": "^7.2.0",
+    "@angular/router": "^7.2.0",
     "big.js": "^5.0.2",
-    "bigi": "^1.4.2",
-    "bip39": "^2.4.0",
-    "bitcoinjs-lib": "^3.3.0",
-    "bs58": "^4.0.1",
-    "core-js": "^2.4.1",
-    "ecurve": "^1.0.5",
+    "bip38": "^2.0.2",
+    "bip39": "^2.5.0",
+    "bitcoinjs-lib": "^5.1.6",
+    "core-js": "^2.5.4",
     "material-design-icons": "^3.0.1",
-    "materialize-css": "1.0.0-alpha.1",
-    "rxjs": "^5.5.6",
-    "wif": "^2.0.6",
-    "zone.js": "^0.8.19"
+    "materialize-css": "^1.0.0",
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "wif": "^3.0.1",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.0",
-    "@angular/compiler-cli": "^5.1.0",
-    "@angular/language-service": "^5.1.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/compiler-cli": "^7.2.0",
+    "@angular/language-service": "^7.2.0",
     "@types/big.js": "^4.0.0",
+    "@types/bip32": "~1.0.0",
     "@types/bip39": "2.4.0",
-    "@types/bitcoinjs-lib": "^3.3.1",
-    "@types/jasmine": "^2.6.0",
-    "@types/jasminewd2": "~2.0.2",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
     "@types/materialize-css": "^1.0.1",
-    "@types/node": "^8.0.46",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "~8.9.4",
+    "@types/wif": "~2.0.1",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
   }
 }
